fix(auth): split Codex metadata helpers out of filestore
Keep filestore.go under the 800-line new-file ratchet after Team
member identity fields were added.
